refactor: added user profile fragment
