+++ common/client/fragments/profile-fragment/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './use-user-state.hook'
+export * from './use-profile-state.hook'
