+++ common/client/fragments/user-profile-fragment/package.json
@@ -1,19 +1,31 @@
 {
-  "name": "@app/profile-fragment",
+  "name": "@app/user-profile-fragment",
   "version": "0.0.1",
   "main": "src/index.ts",
   "dependencies": {
-    "@app/user-profile-fragment": "workspace:0.0.1",
     "@shared/data": "workspace:0.0.1",
+    "@shared/modals-fragment": "workspace:0.0.1",
     "@shared/routes": "workspace:0.0.1",
-    "@ui/condition": "workspace:0.0.1"
+    "@ui/avatar": "workspace:0.0.1",
+    "@ui/button": "workspace:0.0.1",
+    "@ui/card": "workspace:0.0.1",
+    "@ui/condition": "workspace:0.0.1",
+    "@ui/icon": "workspace:0.0.1",
+    "@ui/layout": "workspace:0.0.1",
+    "@ui/link": "workspace:0.0.1",
+    "@ui/tag": "workspace:0.0.1",
+    "@ui/text": "workspace:0.0.1",
+    "@ui/title": "workspace:0.0.1",
+    "@ui/wrapper": "workspace:0.0.1"
   },
   "devDependencies": {
+    "@apollo/client": "3.8.7",
     "@stores/session": "workspace:*",
     "@types/react": "18.2.28",
     "@types/react-dom": "18.2.15",
     "framer-motion": "10.16.1",
-    "next": "12.3.4"
+    "next": "12.3.4",
+    "react-intl": "6.0.8"
   },
   "peerDependencies": {
     "@emotion/css": "*",
