+++ common/client/fragments/user-profile-fragment/src/index.ts
@@ -1,0 +1,2 @@
+export * from './user-profile.component'
+export * from './user-profile.interfaces'
