+++ common/client/fragments/user-profile-fragment/src/profile-avatar/index.ts
