+++ common/client/fragments/user-profile-fragment/src/profile-avatar/profile-avatar.component.tsx
