+++ common/client/fragments/user-profile-fragment/src/profile-avatar/profile-avatar.interfaces.ts
