+++ common/client/fragments/user-profile-fragment/src/profile-description/button-watch-all-skills/button-watch-all-skills.interfaces.ts
