+++ common/client/fragments/user-profile-fragment/src/profile-description/profile-description.interfaces.ts
