+++ common/client/fragments/user-profile-fragment/src/user-profile.interfaces.ts
@@ -1,0 +1,11 @@
+import { Project } from '@shared/data'
+import { Team }    from '@shared/data'
+import { User }    from '@shared/data'
+
+export interface UserProfileProps {
+  user?: User
+  projects: Project[]
+  teams: Team[]
+  isMyProfile: boolean
+  onEditUser: (editedUser: User) => void
+}
